+++ api/src/resolvers/flashCard.ts
@@ -2,10 +2,11 @@ import { IResolvers } from 'graphql-tools'
 
 import { NoteModel, TemplateModel } from '../mongo'
 import { FlashCardDocument } from '../mongo/Note'
+import { globalIdField } from '../utils/graphqlID'
 
 export const root: IResolvers = {
   FlashCard: {
-    id: (root: FlashCardDocument) => root._id.toString(),
+    id: globalIdField(),
     note: (root: FlashCardDocument) => NoteModel.findById(root.noteId),
     template: (root: FlashCardDocument) =>
       TemplateModel.findById(root.templateId),