+++ api/src/resolvers/study.ts
@@ -46,13 +46,14 @@ export const mutations: IResolverObject = {
     ctx: Context
   ) => {
     const { objectId: noteId } = decodeGlobalId(args.noteId)
+    const { objectId: flashCardId } = decodeGlobalId(args.flashCardId)
 
     const note = await NoteModel.findOne({
       _id: noteId,
       ownerId: ctx.user!._id,
     })
 
-    const flashCard = note?.cards.id(args.flashCardId)
+    const flashCard = note?.flashCards.id(flashCardId)
 
     if (!note || !flashCard) {
       throw new ApolloError('FlashCard not found')