+++ app/src/components/__generated__/DecksQuery.ts
@@ -8,28 +8,28 @@
 // ====================================================
 
 export interface DecksQuery_decks {
-  __typename: "Deck";
+  __typename: 'Deck'
   /**
    * Deck id
    */
-  id: string;
+  id: string
   /**
    * Unique identifiable slug
    */
-  slug: string;
+  slug: string
   /**
    * Title of the deck
    */
-  title: string;
+  title: string
   /**
    * Description of the deck
    */
-  description: string | null;
+  description: string | null
 }
 
 export interface DecksQuery {
   /**
    * Retrieve all decks for the logged user
    */
-  decks: DecksQuery_decks[] | null;
+  decks: DecksQuery_decks[]
 }