+++ app/src/utils/noteIdentifier.ts
@@ -2,7 +2,7 @@ import { RawDraftContentState, convertFromRaw } from 'draft-js'
 
 import { NoteQuery_note } from '../components/pages/__generated__/NoteQuery'
 
-type Note = Omit<NoteQuery_note, 'deck' | 'cards' | '__typename'>
+type Note = Omit<NoteQuery_note, 'deck' | 'flashCards' | '__typename'>
 
 export const getNoteIdentifier = (note: Note): string => {
   const primaryFieldValue = note.values.find(